+++ build-release-zips.sh
@@ -42,8 +42,6 @@ for tag in "$@"; do
       --out-dir . \
       --extra-source prefs.js \
       --extra-source VERSION \
-      --extra-source metric-cpu-chip-symbolic.svg \
-      --extra-source metric-memory-dimm-symbolic.svg \
       --extra-source metric-swap-symbolic.svg \
       --extra-source metric-gpu-symbolic.svg \
       --extra-source metric-cpuTemp-symbolic.svg \